Get simple Brick app working
Gonna make a terminal UI using Brick next. It’ll be good to have some kind of UI
for testing as I implement more of the game, and I think structuring things the
way Brick does (event loop that updates state and a function for defining how to
display the state) will give me a good feeling for what state is needed at the
client in order to visualise the game.

Once I’ve got a terminal UI working, I’ll get the game working with two players
and reasonable input, and then I’ll start to look at setting up a server and
client.