Enable making plays and picking up tiles
It’s now possible (and not unpleasant!) to play a few words of Scrabble! The
error feedback is not ideal (error messages are not marked as resolved when the
error condition is fixed), but it does exist.

Next on the list is to enable specifying which letter should be used for a blank
tile. I’m not sure how to do this exactly, I will need some kind of prompt. I
have a vague feeling Brick includes a facility for this – I will investigate.

The weirdest thing I’ve done here is use the Identity monad to write a function
with do notation so I can reuse the same variable name. I think I like it,
though I’m not sure how confusing it would be to someone else.