Split reusable client functionality from brick code
I’d like the current app (where there’s no client-server separation) and the
upcoming terminal client that’ll communicate with the servant server to use the
same client code as much as possible. To do that I’ve split out the
Brick-related code from the more general client code, and defined the client
code in terms of a Backend interface.