Event sourcing!
The state is now stored on the server as two lists of events: a list of
room events and a list of game events. People can join a room, and a
game can be created in a room. In future, multiple games will be allowed
in a room (but maybe only one at a time).

At the moment the client/frontend is still using a “get state” action: I
would like to move to the frontend consuming events from the backend.
Then the refresh becomes “The last event I saw was 96, give me all newer
than that.”. Also we can get minimal communication overhead, hopefully,
and pave the way for using websockets.

Also, the backend currently holds all events in lists, and folds over
the list every time to produce the state. It’d be nicer (I think) to
cache the state (e.g. with projections). But, if I want to provide
events since a certain index, maybe I should use Data.Sequence or
similar? The fact that it’s an append-or-read data structure suggests to
me it should be possible to get some advantages. A vector/array? We’d be
constantly moving it. Hmmmm.