Add haskell-language-server, get working
Unfortunately, I haven’t yet worked out how to do dev dependencies with
nix: I don’t want the language server to be required when building this
for deployment, or running it, but at the moment it is (it’s listed in
the cabal file when it should probably just be in shell.nix). Still,
I’ll get there in time.