WIP: Add draw tiles event
I had initially done the taking of new tiles on the client side, which
is obviously silly. The server needs to do that calculation so that it’s
different for each client and (once I make separate events to be sent to
clients) each client only knows about its own tiles.

The goal now is: do the tile drawing on the server side, so that each
client has a separate rack of tiles, which updates separately.